Moved feature code to accvis.py, begin implementing parser
